File comparison: drop heuristic detection of part size
I was hoping to be able to use the ETag header when multipart was
enabled but we can't be sure of the part size upload. Let's drop it and
only use the Last Modified attribute as soon as either the source object
or destination object has been uploading using multipart. It should
still be enough.